Added beginner expert content/lorem text
The information that I posted about the beginner and expert tracks will
probably stay the same. What will likely change is whether it's
contained within the jumbotron, because right now it seems really large
and over bearing.  As for the content, I think that's probably pretty
good.  Pictures will also be added later on; thinking of doing a 2D and
3D difference kind of thing; maybe Terraria and Minecraft.