Added images; switched to thumbnails
Added images for the beginner and advanced groupings.  Switched to the
thumbnails, because they seem to be a better fit.  The images look a
little too big, so I might scale them down more.  It also could look
better if they were to the right of the text, possibly floating or in
another column.  The caption text could also be more bold.