Images on the beginner page line up better
Used the floating element classes that bootstrap provides, along with
the responsive image class to get images to display in the document in a
way the flows much better. They mostly float right, and it just makes
the content much easier to read. Also made all of the content narrower
by shoving it in column.